Do first Share refactoring
