+++ lib/msf/core/exploit/smb/server/share.rb
@@ -12,6 +12,14 @@ module Msf
   module Exploit::Remote::SMB::Server
     # This mixin provides a minimal SMB server
     module Share
+      require 'msf/core/exploit/smb/server/share/information_level'
+      require 'msf/core/exploit/smb/server/share/sub_command'
+
+      include Msf::Exploit::Remote::SMB::Server::Share::InformationLevel::Find
+      include Msf::Exploit::Remote::SMB::Server::Share::InformationLevel::Query
+      include Msf::Exploit::Remote::SMB::Server::Share::SubCommand::Trans2::FindFirst2
+      include Msf::Exploit::Remote::SMB::Server::Share::SubCommand::Trans2::QueryFileInformation
+      include Msf::Exploit::Remote::SMB::Server::Share::SubCommand::Trans2::QueryPathInformation
 
       include Msf::Exploit::Remote::SMB::Server
 
@@ -388,32 +396,7 @@ module Msf
             smb_cmd_trans_query_file_info_standard(c, buff)
           end
         when CONST::TRANS2_QUERY_PATH_INFO
-          dprint("\t[query_path_info]")
-          loi = ar[144..147].unpack('n*').reverse.pack('n*').to_i(16)
-
-          case mdc # MAX DATA COUNT
-          when CONST::SMB_QUERY_BASIC_MDC
-            case loi
-            when CONST::SMB_QUERY_FILE_BASIC_INFO
-              dprint("\t\t[query_file_info_basic]")
-              smb_cmd_trans_query_file_info_basic(c, buff)
-            else
-              dprint("\t\tUnknown LOI [smb_cmd_trans_query_path_info_basic] - #{loi.to_s}")
-              smb_cmd_trans_query_path_info_basic(c, buff)
-            end
-          when CONST::SMB_QUERY_STANDARD_MDC1, CONST::SMB_QUERY_STANDARD_MDC2
-            dprint("\t\t[query_path_info_standard]")
-            smb_cmd_trans_query_path_info_standard(c, buff)
-          when CONST::SMB_QUERY_FILE_INTERNAL_INFO_MDC
-            dprint("\t\t[query_path_info_standard]")
-            smb_cmd_trans_query_file_info_standard(c, buff)
-          when CONST::SMB_QUERY_FILE_NETWORK_INFO_MDC
-            dprint("\t\t[query_file_info_network]")
-            smb_cmd_trans_query_file_info_network(c, buff)
-          else
-            dprint("\t\tUnknown MDC - Sending to [query_path_info_standard]: #{mdc.to_s}")
-            smb_cmd_trans_query_path_info_standard(c, buff)
-          end
+          smb_cmd_trans2_query_path_information(c, buff)
         when CONST::TRANS2_FIND_FIRST2
           smb_cmd_trans2_find_first2(c, buff)
         else
@@ -426,521 +409,6 @@ module Msf
           pkt['Payload']['SMB'].v['ErrorClass'] = 0xc0000225 # NT_STATUS_NOT_FOUND
           c.put(pkt.to_s)
         end
-      end
-
-      # TODO: delete?
-      def smb_cmd_trans_query_file_info_network(c, buff)
-        dprint("****** [smb_cmd_trans_query_file_info_network] ******")
-        pkt = CONST::SMB_TRANS2_PKT.make_struct
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '').chomp.strip
-
-        if payload.length.to_s.eql?('4')
-          attrib = "\x10\x00\x00\x00" # File attributes => directory
-        else
-          attrib = "\x80\x00\x00\x00" # File attributes => normal file
-        end
-
-        trans2_params = CONST::SMB_TRANS2_QUERY_PATH_INFORMATION_RES_PARAMETERS.make_struct
-        trans2_params.v['EaErrorOffset'] = 0
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-        pkt['Payload']['SMB'].v['Flags1'] = 0x88
-        pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-        pkt['Payload']['SMB'].v['WordCount'] = 10
-        pkt['Payload'].v['ParamCountTotal'] = 2
-        pkt['Payload'].v['DataCountTotal'] = 56
-        pkt['Payload'].v['ParamCount'] = 2
-        pkt['Payload'].v['ParamOffset'] = 56
-        pkt['Payload'].v['DataCount'] = 56
-        pkt['Payload'].v['DataOffset'] = 60
-        pkt['Payload'].v['Payload'] =
-          "\x00" + # Padding
-          trans2_params.to_s +
-          "\x00\x00" + # Padding
-          # QUERY_PATH_INFO Data
-          [lo, hi].pack("VV") + # Created
-          [lo, hi].pack("VV") + # Last Access
-          [lo, hi].pack("VV") + # Last Write
-          [lo, hi].pack("VV") + # Change
-          "\x00\x00\x10\x00\x00\x00\x00\x00" + # Allocation Size = 1048576 || 1Mb
-          [exe_contents.length].pack("V") + "\x00\x00\x00\x00" + # End Of File
-          attrib +
-          "\x00\x00\x00\x00" # Unknown
-        c.put(pkt.to_s)
-      end
-
-      # TODO: delete?
-      def smb_cmd_trans_query_file_info_standard(c, buff)
-        dprint("****** [smb_cmd_trans_query_file_info_standard] ****** ")
-
-        trans2_params = CONST::SMB_TRANS2_QUERY_PATH_INFORMATION_RES_PARAMETERS.make_struct
-        trans2_params.v['EaErrorOffset'] = 0
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-        pkt['Payload']['SMB'].v['Flags1'] = 0x88
-        pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-        pkt['Payload']['SMB'].v['WordCount'] = 10
-        pkt['Payload'].v['ParamCountTotal'] = 2
-        pkt['Payload'].v['DataCountTotal'] = 8
-        pkt['Payload'].v['ParamCount'] = 2
-        pkt['Payload'].v['ParamOffset'] = 56
-        pkt['Payload'].v['DataCount'] = 8
-        pkt['Payload'].v['DataOffset'] = 60
-        pkt['Payload'].v['Payload'] =
-        "\x00" + # Padding
-        # QUERY_FILE Parameters
-        trans2_params.to_s +
-        "\x00\x00" + # Padding
-        # QUERY_FILE_INFO Data
-        "\x95\x1c\x02\x00\x00\x00\x00\x00"
-        c.put(pkt.to_s)
-      end
-
-      #
-      # Responds to QUERY_PATH_INFO (Standard) requests
-      #
-      def smb_cmd_trans_query_path_info_standard(c, buff)
-        dprint("[smb_cmd_trans_query_path_info_standard]")
-        smb = @state[c]
-        pkt = CONST::SMB_TRANS2_PKT.make_struct
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '').chomp.strip
-        ar = Rex::Text.to_hex(buff, '').to_s
-        fid = ar[144..147].unpack('n*').reverse.pack('n*')
-
-        # If FileID matches or matches file, send file response
-        if ( fid.hex.eql?(smb[:file_id].to_i) or payload.length.eql?(file_name.length) )
-          attrib2 = 0 # IsFile
-        else
-          # Otherwise return a Directory answer
-          attrib2 = 1 # IsDir
-        end
-
-        trans2_params = CONST::SMB_TRANS2_QUERY_PATH_INFORMATION_RES_PARAMETERS.make_struct
-        trans2_params.v['EaErrorOffset'] = 0
-
-        query_path_info = CONST::SMB_QUERY_FILE_STANDARD_INFO_HDR.make_struct
-        query_path_info.v['AllocationSize'] = 1048576
-        query_path_info.v['EndOfFile'] = exe_contents.length
-        query_path_info.v['NumberOfLinks'] = 1
-        query_path_info.v['DeletePending'] = 0
-        query_path_info.v['Directory'] = attrib2
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-        pkt['Payload']['SMB'].v['Flags1'] = 0x88
-        pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-        pkt['Payload']['SMB'].v['WordCount'] = 10
-        pkt['Payload'].v['ParamCountTotal'] = 2
-        pkt['Payload'].v['DataCountTotal'] = 24
-        pkt['Payload'].v['ParamCount'] = 2
-        pkt['Payload'].v['ParamOffset'] = 56
-        pkt['Payload'].v['DataCount'] = 24
-        pkt['Payload'].v['DataOffset'] = 60
-        pkt['Payload'].v['Payload'] =
-          "\x00" + # Padding
-          trans2_params.to_s +
-          "\x00\x00" + # Padding
-          query_path_info.to_s +
-          "\x00\x00" # Unknown
-        c.put(pkt.to_s)
-      end
-
-      #
-      # Responds to QUERY_FILE_INFO (Basic) requests
-      #
-      def smb_cmd_trans_query_file_info_basic(c, buff)
-        dprint("[smb_cmd_trans_query_file_info_basic]")
-        pkt = CONST::SMB_TRANS2_PKT.make_struct
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '\\').chomp.strip
-        file = file_name
-        path = path_name
-
-        begin
-          fileext = file.split('.').last
-        rescue
-          fileext = file
-        end
-
-        begin
-          payext = payload.split('.').last
-        rescue
-          payext = payload
-        end
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        # If payload contains our file extension, send file response
-        if payext and payext.downcase.eql?(fileext.downcase)
-          attrib = SMB_EXT_FILE_ATTR_NORMAL # File attributes => file
-        elsif payload.length.to_s.eql?('1') or payload.eql?(path)
-          # if QUERY_PATH_INFO_PARAMETERS doesn't include a file name,
-          # return a Directory answer
-          attrib = CONST::SMB_EXT_FILE_ATTR_DIRECTORY # File attributes => directory
-        end
-
-        if (payext and payext.downcase.eql?(fileext.downcase)) or payload.length.to_s.eql?('1') or payload.length.to_s.eql?('4') or payload.eql?(path)
-
-          trans2_params = CONST::SMB_TRANS2_QUERY_PATH_INFORMATION_RES_PARAMETERS.make_struct
-          trans2_params.v['EaErrorOffset'] = 0
-
-          query_path_info = CONST::SMB_QUERY_FILE_BASIC_INFO_HDR.make_struct
-          query_path_info.v['loCreationTime'] = lo
-          query_path_info.v['hiCreationTime'] = hi
-          query_path_info.v['loLastAccessTime'] = lo
-          query_path_info.v['hiLastAccessTime'] = hi
-          query_path_info.v['loLastWriteTime'] = lo
-          query_path_info.v['hiLastWriteTime'] = hi
-          query_path_info.v['loLastChangeTime'] = lo
-          query_path_info.v['hiLastChangeTime'] = hi
-          query_path_info.v['ExtFileAttributes'] = attrib
-
-          pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-          pkt['Payload']['SMB'].v['Flags1'] = 0x88
-          pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-          pkt['Payload']['SMB'].v['WordCount'] = 10
-          pkt['Payload'].v['ParamCountTotal'] = 2
-          pkt['Payload'].v['DataCountTotal'] = 40
-          pkt['Payload'].v['ParamCount'] = 2
-          pkt['Payload'].v['ParamOffset'] = 56
-          pkt['Payload'].v['DataCount'] = 40
-          pkt['Payload'].v['DataOffset'] = 60
-          pkt['Payload'].v['Payload'] =
-            "\x00" + # Padding
-            trans2_params.to_s +
-            "\x00\x00" + # Padding
-            query_path_info.to_s
-          c.put(pkt.to_s)
-        else
-          pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-          pkt['Payload']['SMB'].v['ErrorClass'] = 0xC0000034 # OBJECT_NAME_NOT_FOUND
-          pkt['Payload']['SMB'].v['Flags1'] = 0x88
-          pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-          c.put(pkt.to_s)
-        end
-      end
-
-      #
-      # Responds to QUERY_PATH_INFO (Basic) requests
-      #
-      def smb_cmd_trans_query_path_info_basic(c, buff)
-        dprint("[smb_cmd_trans_query_path_info_basic]")
-        pkt = CONST::SMB_TRANS2_PKT.make_struct
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '\\').chomp.strip.split('\\').last
-        file = file_name
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        # If payload contains our file, send a file response
-        # For MS13-071, payload is an empty unicode request (4 bytes)
-        if payload.downcase.eql?(file.downcase) or payload.length.to_s.eql?('4')
-          if payload.length.to_s.eql?('4')
-            attrib = CONST::SMB_EXT_FILE_ATTR_DIRECTORY # File attributes => directory
-          else
-            attrib = CONST::SMB_EXT_FILE_ATTR_NORMAL # File attributes => file
-          end
-        else
-          if payload.downcase.include?(file.downcase)
-            attrib = CONST::SMB_EXT_FILE_ATTR_NORMAL # File attributes => file
-          else
-            # QUERY_PATH_INFO_PARAMETERS doesn't include a file name, return a Directory answer
-            attrib = CONST::SMB_EXT_FILE_ATTR_DIRECTORY # File attributes => directory
-          end
-        end
-
-        trans2_params = CONST::SMB_TRANS2_QUERY_PATH_INFORMATION_RES_PARAMETERS.make_struct
-        trans2_params.v['EaErrorOffset'] = 0
-
-        query_path_info = CONST::SMB_QUERY_FILE_BASIC_INFO_HDR.make_struct
-        query_path_info.v['loCreationTime'] = lo
-        query_path_info.v['hiCreationTime'] = hi
-        query_path_info.v['loLastAccessTime'] = lo
-        query_path_info.v['hiLastAccessTime'] = hi
-        query_path_info.v['loLastWriteTime'] = lo
-        query_path_info.v['hiLastWriteTime'] = hi
-        query_path_info.v['loLastChangeTime'] = lo
-        query_path_info.v['hiLastChangeTime'] = hi
-        query_path_info.v['ExtFileAttributes'] = attrib
-
-        # If payload contains our file, send the response
-        if payload.downcase.eql?(file.downcase) or payload.length.to_s.eql?('4') or payload.downcase.include?(file.downcase)
-          pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-          pkt['Payload']['SMB'].v['Flags1'] = 0x88
-          pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-          pkt['Payload']['SMB'].v['WordCount'] = 10
-          pkt['Payload'].v['ParamCountTotal'] = 2
-          pkt['Payload'].v['DataCountTotal'] = 40
-          pkt['Payload'].v['ParamCount'] = 2
-          pkt['Payload'].v['ParamOffset'] = 56
-          pkt['Payload'].v['DataCount'] = 40
-          pkt['Payload'].v['DataOffset'] = 60
-          pkt['Payload'].v['Payload'] =
-            "\x00" + # Padding
-            trans2_params.to_s +
-            "\x00\x00" + # Padding
-            query_path_info.to_s
-          c.put(pkt.to_s)
-        else
-          # Else send not found
-          pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-          pkt['Payload']['SMB'].v['ErrorClass'] = 0xC0000034 # OBJECT_NAME_NOT_FOUND
-          pkt['Payload']['SMB'].v['Flags1'] = 0x88
-          pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-          c.put(pkt.to_s)
-        end
-      end
-
-      def smb_cmd_trans2_find_first2(c, buff)
-        dprint("[smb_cmd_trans2_find_first2]")
-
-        ar = Rex::Text.to_hex(buff, '').to_s
-        loi = ar[156..159].unpack('n*').reverse.pack('n*').to_i(16)
-
-        case loi
-        when CONST::SMB_FIND_FILE_NAMES_INFO
-          dprint("\t[SMB_FIND_FILE_NAMES_INFO]")
-          smb_cmd_find_file_names_info(c, buff)
-        when CONST::SMB_FIND_FILE_BOTH_DIRECTORY_INFO
-          dprint("\t[SMB_FIND_FILE_BOTH_DIRECTORY_INFO]")
-          smb_cmd_find_file_both_directory_info(c, buff)
-        when CONST::SMB_FIND_FILE_FULL_DIRECTORY_INFO
-          dprint("\t[SMB_FIND_FILE_FULL_DIRECTORY_INFO]")
-          smb_cmd_find_file_full_directory_info(c, buff)
-        else
-          dprint("\t[Unexpected :?]") #TODO
-          #smb_cmd_trans_find_first2(c, buff)
-          fail_with(Failure::Unknown, "Invalid or unknown TRANS2_FIND_FIRST2 REQUEST???") #TODO
-        end
-      end
-
-      #
-      # Responds to FIND_FIRST2 requests
-      # Command: Find File Both Directory Info
-      #
-      def smb_cmd_find_file_both_directory_info(c, buff)
-        dprint("[smb_cmd_trans_find_first2]")
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '\\').chomp.strip
-
-        print_status("#{Rex::Text.to_hex_dump(payload)}")
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        if payload && payload.include?(file_name)
-          data = Rex::Text.to_unicode(file_name)
-          length = exe_contents.length
-          ea = 0
-          alloc = 1048576 # Allocation Size = 1048576 || 1Mb
-          attrib = CONST::SMB_EXT_FILE_ATTR_NORMAL
-          search = 1
-        elsif payload && payload == path_name
-          data = Rex::Text.to_unicode(path)
-          length = 0
-          ea = 0x21
-          alloc = 0 # 0Mb
-          attrib = CONST::SMB_EXT_FILE_ATTR_DIRECTORY
-          pkt['Payload'].v['SetupCount'] = 0
-          search = 0x100
-        else
-          smb_error(CONST::SMB_COM_TRANSACTION2, c, CONST::SMB_STATUS_NO_SUCH_FILE, true)
-          return
-        end
-
-        trans2_params = CONST::SMB_TRANS2_FIND_FIRST2_RES_PARAMETERS.make_struct
-        trans2_params.v['SID'] = 0xfffd
-        trans2_params.v['SearchCount'] = search
-        trans2_params.v['EndOfSearch'] = search
-        trans2_params.v['EaErrorOffset'] = ea
-        trans2_params.v['LastNameOffset'] = 0
-
-        find_file = CONST::SMB_FIND_FILE_BOTH_DIRECTORY_INFO_HDR.make_struct
-        find_file.v['NextEntryOffset'] = CONST::SMB_FIND_FILE_BOTH_DIRECTORY_INFO_HDR_LENGTH + data.length
-        find_file.v['FileIndex'] = 0
-        find_file.v['loCreationTime'] = lo
-        find_file.v['hiCreationTime'] = hi
-        find_file.v['loLastAccessTime'] = lo
-        find_file.v['hiLastAccessTime'] = hi
-        find_file.v['loLastWriteTime'] = lo
-        find_file.v['hiLastWriteTime'] = hi
-        find_file.v['loLastChangeTime'] = lo
-        find_file.v['hiLastChangeTime'] = hi
-        find_file.v['EndOfFile'] = length
-        find_file.v['AllocationSize'] = alloc
-        find_file.v['ExtFileAttributes'] = attrib
-        find_file.v['FileName'] = data
-
-        pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-        pkt['Payload']['SMB'].v['Flags1'] = 0x88
-        pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-        pkt['Payload']['SMB'].v['WordCount'] = 10
-        pkt['Payload'].v['ParamCountTotal'] = 10
-        pkt['Payload'].v['DataCountTotal'] = 14 + data.length
-        pkt['Payload'].v['ParamCount'] = 10
-        pkt['Payload'].v['ParamOffset'] = 56
-        pkt['Payload'].v['DataCount'] = 14 + data.length
-        pkt['Payload'].v['DataOffset'] = 68
-        pkt['Payload'].v['Payload'] =
-          "\x00" + # Padding
-          trans2_params.to_s +
-          "\x00\x00" + # Padding
-          find_file.to_s
-        c.put(pkt.to_s)
-      end
-
-      #
-      # Responds to FIND_FIRST2 requests
-      # Command: Find File Names Info
-      #
-      def smb_cmd_find_file_names_info(c, buff)
-        dprint("[smb_cmd_trans_find_first2_file]")
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '\\').chomp.strip.split('\\').last
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        if payload && payload.include?(file_name)
-          data = Rex::Text.to_unicode(file_name)
-        elsif payload && payload == path_name
-          data = Rex::Text.to_unicode(path_name)
-        else
-          smb_error(CONST::SMB_COM_TRANSACTION2, c, CONST::SMB_STATUS_NO_SUCH_FILE, true)
-          return
-        end
-
-        find_file = CONST::SMB_FIND_FILE_NAMES_INFO_HDR.make_struct
-        find_file.v['NextEntryOffset'] = CONST::SMB_FIND_FILE_NAMES_INFO_HDR_LENGTH + data.length
-        find_file.v['FileIndex'] = 0
-        find_file.v['FileName'] = data
-
-        trans2_params = CONST::SMB_TRANS2_FIND_FIRST2_RES_PARAMETERS.make_struct
-        trans2_params.v['SID'] = 0xfffd
-        trans2_params.v['SearchCount'] = 1
-        trans2_params.v['EndOfSearch'] = 1
-        trans2_params.v['EaErrorOffset'] = 0
-        trans2_params.v['LastNameOffset'] = 0
-
-        # If its asking for a file, return file
-        pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-        pkt['Payload']['SMB'].v['Flags1'] = 0x88
-        pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-        pkt['Payload']['SMB'].v['WordCount'] = 10
-        pkt['Payload'].v['ParamCountTotal'] = 10
-        pkt['Payload'].v['DataCountTotal'] = 14 + data.length
-        pkt['Payload'].v['ParamCount'] = 10
-        pkt['Payload'].v['ParamOffset'] = 56
-        pkt['Payload'].v['DataCount'] = 14 + data.length
-        pkt['Payload'].v['DataOffset'] = 68
-        pkt['Payload'].v['Payload'] =
-          "\x00" + # Padding
-          trans2_params.to_s +
-          "\x00\x00" + # Padding
-          find_file.to_s
-        c.put(pkt.to_s)
-      end
-
-      #
-      # Responds to FIND_FIRST2 requests
-      # Command: Find File Full Directory Info
-      #
-      def smb_cmd_find_file_full_directory_info(c, buff)
-        dprint("[smb_cmd_trans_find_first2_full]")
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        pkt.from_s(buff)
-
-        payload = pkt['Payload'].v['SetupData'].gsub(/\x00/, '').gsub(/.*\\/, '\\').chomp.strip
-
-        pkt = CONST::SMB_TRANS_RES_PKT.make_struct
-        smb_set_defaults(c, pkt)
-
-        if payload && payload.include?(file_name)
-          data = Rex::Text.to_unicode(file_name)
-          length = exe_contents.length
-          ea = 0
-          alloc = 1048576 # Allocation Size = 1048576 || 1Mb
-          attrib = CONST::SMB_EXT_FILE_ATTR_NORMAL # File
-          search = 0x100
-        elsif payload && payload == path_name
-          data = path
-          length = 0
-          ea = 0x21
-          alloc = 0 # 0Mb
-          attrib = CONST::SMB_EXT_FILE_ATTR_DIRECTORY # Dir
-          pkt['Payload'].v['SetupCount'] = 0
-          search = 1
-        else
-          smb_error(CONST::SMB_COM_TRANSACTION2, c, CONST::SMB_STATUS_NO_SUCH_FILE, true)
-          return
-        end
-
-        find_file = CONST::SMB_FIND_FILE_FULL_DIRECTORY_INFO_HDR.make_struct
-        find_file.v['NextEntryOffset'] = CONST::SMB_FIND_FILE_FULL_DIRECTORY_INFO_HDR_LENGTH + data.length
-        find_file.v['FileIndex'] = 0
-        find_file.v['loCreationTime'] = lo
-        find_file.v['hiCreationTime'] = hi
-        find_file.v['loLastAccessTime'] = lo
-        find_file.v['hiLastAccessTime'] = hi
-        find_file.v['loLastWriteTime'] = lo
-        find_file.v['hiLastWriteTime'] = hi
-        find_file.v['loLastChangeTime'] = lo
-        find_file.v['hiLastChangeTime'] = hi
-        find_file.v['EndOfFile'] = length
-        find_file.v['AllocationSize'] = alloc
-        find_file.v['ExtFileAttributes'] = attrib
-        find_file.v['FileName'] = data
-
-        trans2_params = CONST::SMB_TRANS2_FIND_FIRST2_RES_PARAMETERS.make_struct
-        trans2_params.v['SID'] = 0xfffd
-        trans2_params.v['SearchCount'] = search
-        trans2_params.v['EndOfSearch'] = search
-        trans2_params.v['EaErrorOffset'] = ea
-        trans2_params.v['LastNameOffset'] = 0
-
-        pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_TRANSACTION2
-        pkt['Payload']['SMB'].v['Flags1'] = 0x88
-        pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
-        pkt['Payload']['SMB'].v['WordCount'] = 10
-        pkt['Payload'].v['ParamCountTotal'] = 10
-        pkt['Payload'].v['DataCountTotal'] = 68 + data.length
-        pkt['Payload'].v['ParamCount'] = 10
-        pkt['Payload'].v['ParamOffset'] = 56
-        pkt['Payload'].v['DataCount'] = 68 + data.length
-        pkt['Payload'].v['DataOffset'] = 68
-        pkt['Payload'].v['Payload'] =
-          "\x00" + # Padding
-          trans2_params.to_s +
-          "\x00\x00" + # Padding
-          find_file.to_s
-
-        c.put(pkt.to_s)
       end
     end
   end
