+++ lib/msf/core/exploit/smb/server/share/information_level.rb
@@ -1,0 +1,12 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit::Remote::SMB::Server
+    module Share
+      module InformationLevel
+        require 'msf/core/exploit/smb/server/share/information_level/find'
+        require 'msf/core/exploit/smb/server/share/information_level/query'
+      end
+    end
+  end
+end
