+++ lib/msf/core/exploit/smb/server/share/sub_command.rb
@@ -1,0 +1,11 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit::Remote::SMB::Server
+    module Share
+      module SubCommand
+        require 'msf/core/exploit/smb/server/share/sub_command/trans2'
+      end
+    end
+  end
+end
