+++ lib/msf/core/exploit/smb/server/share/sub_command/trans2.rb
@@ -1,0 +1,15 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit::Remote::SMB::Server
+    module Share
+      module SubCommand
+        module Trans2
+          require 'msf/core/exploit/smb/server/share/sub_command/trans2/find_first2'
+          require 'msf/core/exploit/smb/server/share/sub_command/trans2/query_file_information'
+          require 'msf/core/exploit/smb/server/share/sub_command/trans2/query_path_information'
+        end
+      end
+    end
+  end
+end
