+++ lib/msf/core/exploit/smb/server/share/sub_command/trans2/query_file_information.rb
@@ -1,0 +1,45 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit::Remote::SMB::Server
+    module Share
+      module SubCommand
+        module Trans2
+          # This mixin provides methods to handle TRAN2_QUERY_PATH_INFORMATION subcommands
+          module QueryFileInformation
+            def smb_cmd_trans2_query_path_information(c, buff)
+              dprint("[smb_cmd_trans2_query_path_information]")
+              ar = Rex::Text.to_hex(buff, '').to_s
+              mdc = ar[86..89].unpack('n*').reverse.pack('n*').to_i(16)
+              loi = ar[144..147].unpack('n*').reverse.pack('n*').to_i(16)
+
+              case mdc # MAX DATA COUNT
+              when CONST::SMB_QUERY_BASIC_MDC
+                case loi
+                when CONST::SMB_QUERY_FILE_BASIC_INFO
+                  dprint("\t\t[query_file_info_basic]")
+                  smb_cmd_trans_query_file_info_basic(c, buff)
+                else
+                  dprint("\t\tUnknown LOI [smb_cmd_trans_query_path_info_basic] - #{loi.to_s}")
+                  smb_cmd_trans_query_path_info_basic(c, buff)
+                end
+              when CONST::SMB_QUERY_STANDARD_MDC1, CONST::SMB_QUERY_STANDARD_MDC2
+                dprint("\t\t[query_path_info_standard]")
+                smb_cmd_trans_query_path_info_standard(c, buff)
+              when CONST::SMB_QUERY_FILE_INTERNAL_INFO_MDC
+                dprint("\t\t[query_path_info_standard]")
+                smb_cmd_trans_query_file_info_standard(c, buff)
+              when CONST::SMB_QUERY_FILE_NETWORK_INFO_MDC
+                dprint("\t\t[query_file_info_network]")
+                smb_cmd_trans_query_file_info_network(c, buff)
+              else
+                dprint("\t\tUnknown MDC - Sending to [query_path_info_standard]: #{mdc.to_s}")
+                smb_cmd_trans_query_path_info_standard(c, buff)
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
